requirements.txt
kafka-python
requests
websocket-client
rel
